Move generatePDF into a module
